move to summary tab, and dummy pv tab
