+++ application/tabs/summary/callbacks.py
@@ -2,7 +2,7 @@
 from typing import List
 
 from dash import Input, Output, State
-from plots import make_plot
+from .plots import make_plot
 
 
 def make_callbacks(app):
