+++ application/tabs/summary/plots.py
